Update webpack and the store
