+++ src/js/components/InfoStatus/InfoStatus.spec.js
