+++ src/js/components/InfoStatus/index.js
