+++ src/js/components/Loader/Loader.spec.jsx
