+++ src/js/components/LoadingBar/LoadingBar.jsx
