+++ src/js/components/LoadingBar/LoadingBar.spec.jsx
