+++ src/js/components/LoadingBar/index.js
