+++ src/js/components/LoadingSpinner/index.jsx
