+++ src/js/components/PrivateRoute/PrivateRoute.spec.jsx
