+++ src/js/components/StatusBar/StatusBar.spec.jsx
