+++ src/js/components/StatusBar/index.jsx
