+++ src/js/store/index.js
@@ -1,5 +1,34 @@
-if (process.env.NODE_ENV === 'production') {
-  module.exports = require('./prod'); // eslint-disable-line global-require
-} else {
-  module.exports = require('./dev'); // eslint-disable-line global-require
-}
+import { createStore, applyMiddleware, compose } from 'redux';
+import createSagaMiddleware from 'redux-saga';
+import { connectRouter, routerMiddleware } from 'connected-react-router';
+import history from 'history';
+import reducer from 'store/reducers';
+import status from 'store/status/middleware';
+import rootSaga from 'store/sagas';
+
+export default () => {
+  const sagaMiddleware = createSagaMiddleware();
+  // eslint-disable-next-line no-underscore-dangle
+  const composeEnhancers = (__DEV__ && window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__) || compose;
+  const store = createStore(
+    connectRouter(history)(reducer),
+    composeEnhancers(
+      applyMiddleware(
+        status,
+        routerMiddleware(history),
+        sagaMiddleware,
+      ),
+    ),
+  );
+
+  if (module.hot && __DEV__) {
+    module.hot.accept('store/reducers', () => {
+      const nextReducer = require('store/reducers').default; // eslint-disable-line global-require
+      store.replaceReducer(nextReducer);
+    });
+  }
+
+  sagaMiddleware.run(rootSaga);
+
+  return store;
+};
